requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.4.0
Flask_QRcode==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
python-dotenv
